api_yamdb/requirements.txt
asgiref==3.2.10
gunicorn==20.0.4
psycopg2-binary==2.8.6
atomicwrites==1.4.1
attrs==22.1.0
black==22.8.0
certifi==2022.6.15.2
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-yasg==1.21.3
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==0.13.1
py==1.11.0
PyJWT==2.1.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2022.2.1
requests==2.26.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.12
